rename das service files
